Eliminate the mocha-eslint dependency
The mocha-eslint package appears abandoned.  It pulls in an older version
of eslint (^4.2.0, currently resolving to 4.19.1) than is required by
eslint-config-wikimedia (^5.9.0).  This is true even for the current mocha-
eslint version.

This patch bumps the requested eslint version to ^5.9.0 and drops the
mocha-eslint dependency in favor of a direct eslint invocation in the npm
test script. One minor spacing change was needed to comply with the update.

Upstreams the related mobileapps change:
https://gerrit.wikimedia.org/r/#/c/mediawiki/services/mobileapps/+/475832/

Bug: https://phabricator.wikimedia.org/T210460