Propagate the request ID when issuing requests; rel v0.7.0
This commit changes the API of the Request object and the utility
functions used to contact the MW and REST APIs. The Request object now
has a `context` property that includes the request ID. Additionally, the
request now features an `issueRequest()` method which should be used
instead of `preq`. It augments the request by setting the user-agent and
x-request-id headers automatically, in this way propagating the request
ID further down the call chain.

In order to account for these changes, `mwApiGet()` and `restApiGet()`
have been changed to take the request as a parameter instead of the
application object.

Bug: T225711