Harden default CSP settings within app.js
The Service-Template-Node code is used for many internal
wikimedia services where security headers are likely unnecessary.
Still, if defaults are to be provided, those should be as secure
and up-to-date as possible.  Especially if any service were to
ever be made public in any way or an attacker found some novel
way of exploiting various services lacking sufficient protection
via security headers.  This pull request:

* Hardens default CSP policy, namely adding values of 'none' for
various sources and specifying the base-uri directive.

* Removes the deprecated x-webkit-csp and x-content-security-policy
headers.

Note: per developer.mozilla.org docs, x-frame-options is
obsolete if a frame-ancestors directive is specified within the
CSP.  x-xss-protection is also obsolete with a CSP that disallows
unsafe-inline scripts.  Both of these are satisfied by STN and
can be removed.  access-control-allow-origin should likely not
default to '*', but I don't have a good suggestion of an
appropriate default value that wouldn't potentially break
things somewhat unnecessarily.

Bug: T262652